chore: split internal/utils grab-bag into focused packages
