+++ internal/prometheus/prometheus.go
@@ -1,4 +1,4 @@
-package utils
+package prometheus
 
 import (
 	"context"
