+++ internal/prometheus/prometheus_test.go
@@ -1,4 +1,4 @@
-package utils
+package prometheus
 
 import (
 	"context"
