+++ internal/prometheus/prometheus_transport.go
@@ -1,4 +1,4 @@
-package utils
+package prometheus
 
 import (
 	"fmt"
