+++ internal/prometheus/tls.go
@@ -1,4 +1,4 @@
-package utils
+package prometheus
 
 import (
 	"crypto/tls"
