+++ internal/prometheus/tls_test.go
@@ -1,4 +1,4 @@
-package utils
+package prometheus
 
 import (
 	"crypto/tls"
